Improve CodeClimate
